requirements.txt
python-telegram-bot
python-dotenv
reportlab
Pillow